Perfected global updater procedure

+++ application/controllers/internal/Autoexec_updater.php
@@ -1,6 +1,7 @@
 <?php
 ini_set('mysql.connect_timeout', 3000);
 ini_set('default_socket_timeout', 3000);
+ini_set('max_execution_time', 0);
 
 defined('BASEPATH') or exit('No direct script access allowed');
 
@@ -14,6 +15,7 @@ class Autoexec_updater extends CI_Controller
         $this->db->cache_off();
         $this->db->cache_delete_all();
         $this->load->model('common/Log');
+        $this->load->model('common/Msg');
         $this->load->model('internal/Autoexec_updater_model');
         $this->load->model('Updater_model');
     }
@@ -23,60 +25,75 @@ class Autoexec_updater extends CI_Controller
         $chars = $this->Autoexec_updater_model->getAllUsers();
 
         foreach ($chars as $row) {
-            $start         = microtime(true);
+            $start    = microtime(true);
             $username = $row->username;
             echo $username . "\n";
 
             $this->Updater_model->init($username);
 
             if (!$this->Updater_model->testEndpoint()) {
-                echo Msg::XML_CONNECT_FAILURE;
+                echo XML_CONNECT_FAILURE;
             } else if (!$this->Updater_model->processAPIKeys($username)) {
-                echo Msg::XML_CONNECT_FAILURE;
+                echo XML_CONNECT_FAILURE;
             } else {
 
-                try {
-                    //catch the API acess violation bug
-                    $result_iterate = $this->Updater_model->iterateAccountCharacters();
-                    if ($result_iterate == "dberror") {
-                        //in case the transaction fails
-                        echo Msg::DB_ERROR;
+                if ($this->Updater_model->isLocked($username)) {
+                    log_message('error', $username . ' is locked');
+                } else {
+                    try {
+                        $this->Updater_model->lock($username);
+                        log_message('error', $username . ' locked initial');
+                        //catch the API acess violation bug
+                        $result_iterate = $this->Updater_model->iterateAccountCharacters();
+
+                        if ($result_iterate == "noChars") {
+                            echo Msg::LOGIN_NO_CHARS;
+
+                        } else if ($result_iterate == "dberror") {
+                            echo Msg::DB_ERROR;
+                            return;
+                        }
+                    } catch (\Pheal\Exceptions\PhealException $e) {
+                        //in case the API throws an exception (usually a bug)
+                        echo sprintf(
+                            "an exception was caught! Type: %s Message: %s",
+                            get_class($e),
+                            $e->getMessage()
+                        );
+
+                        //remove cache and try again
+                        $problematicKeys = $this->Updater_model->getAPIKeys($this->session->iduser);
+
+                        foreach ($problematicKeys as $row) {
+                            $key = $row->key;
+                            $dir = FILESTORAGE . $key;
+                            $this->removeDirectory($path);
+                            $this->Log->addEntry('clear', $this->session->iduser);
+
+                            $this->Updater_model->release($username);
+                            log_message('error', $username . ' released errpr');
+                        }
                         return;
                     }
-                } catch (\Pheal\Exceptions\PhealException $e) {
-                    //in case the API throws an exception
-                    echo sprintf(
-                        "an exception was caught! Type: %s Message: %s",
-                        get_class($e),
-                        $e->getMessage()
-                    );
-
-                    $problematicKeys = $this->Updater_model->getAPIKeys($username);
-
-                    foreach ($problematicKeys as $row2) {
-                        $key = $row2->key;
-                        $dir = FILESTORAGE . $key;
-                        $this->removeDirectory($path);
-                        $this->Log->addEntry('clear', $username);
-                    }
-                    return;
                 }
 
                 //calculate profits
                 $this->db->trans_start();
                 $this->Updater_model->calculateProfits();
-                //iterate
+                //totals and history
                 $this->Updater_model->updateTotals();
                 $this->db->trans_complete();
 
                 if ($this->db->trans_status() === false) {
                     echo Msg::DB_ERROR;
                     return;
+                } else {
+                    $finish = microtime(true);
+                    echo number_format($finish - $start, 2) . " for username " . $username . "\n";
+                    $this->Updater_model->release($username);
+                    log_message('error', $username .' released');
                 }
             }
-
-            $finish = microtime(true);
-            echo number_format($finish - $start, 2) . " for username " . $username . "\n";
         }
     }
